+++ backend/src/Webhook/Connector/Mastodon.php
@@ -15,14 +15,6 @@ use App\Utilities\Urls;
  */
 final class Mastodon extends AbstractSocialConnector
 {
-    protected function getRateLimitTime(StationWebhook $webhook): int
-    {
-        $config = $webhook->getConfig();
-        $rateLimitSeconds = Types::int($config['rate_limit'] ?? null);
-
-        return max(10, $rateLimitSeconds);
-    }
-
     public function dispatch(
         Station $station,
         StationWebhook $webhook,